+++ theHarvester/lib/core.py
@@ -416,7 +416,6 @@ class Core:
             'dehashed',
             'dnsdb',
             'dnsdumpster',
-            'duckduckgo',
             'dymo',
             'fofa',
             'fullhunt',